+++ modules/core/src/main/java/io/github/palexdev/mfxcore/collections/CircularQueue.java
@@ -29,38 +29,14 @@ public class CircularQueue<E> extends LinkedList<E> {
 	//================================================================================
 	// Properties
 	//================================================================================
-	private int size;
+	private int capacity;
 
 	//================================================================================
 	// Constructors
 	//================================================================================
-	public CircularQueue(int size) {
+	public CircularQueue(int capacity) {
 		super();
-		this.size = size;
-	}
-
-	//================================================================================
-	// Methods
-	//================================================================================
-
-	/**
-	 * Sets the maximum size of the queue and removes exceeding elements
-	 * if the specified size is lesser than the number of elements.
-	 *
-	 * @param size The new desired size
-	 * @throws IllegalArgumentException if the desired size is 0
-	 */
-	public void setSize(int size) {
-		if (size == 0) {
-			throw new IllegalArgumentException("Size cannot be 0!");
-		}
-
-		if (size < super.size()) {
-			for (int i = 0; i < (super.size() - size); i++) {
-				super.remove();
-			}
-		}
-		this.size = size;
+		this.capacity = capacity;
 	}
 
 	//================================================================================
@@ -75,9 +51,41 @@ public class CircularQueue<E> extends LinkedList<E> {
 	 */
 	@Override
 	public boolean add(E e) {
-		if (super.size() == this.size) {
+		if (super.size() == this.capacity) {
 			super.remove();
 		}
 		return super.add(e);
+	}
+
+	//================================================================================
+	// Getters/Setters
+	//================================================================================
+
+	/**
+	 * @return the maximum number of elements the queue can contain before it starts deleting them
+	 * from the head
+	 */
+	public int getCapacity() {
+		return capacity;
+	}
+
+	/**
+	 * Sets the maximum size of the queue and removes exceeding elements
+	 * if the specified size is lesser than the number of elements.
+	 *
+	 * @param capacity The new desired size
+	 * @throws IllegalArgumentException if the desired size is 0
+	 */
+	public void setCapacity(int capacity) {
+		if (capacity == 0) {
+			throw new IllegalArgumentException("Size cannot be 0!");
+		}
+
+		if (capacity < super.size()) {
+			for (int i = 0; i < (super.size() - capacity); i++) {
+				super.remove();
+			}
+		}
+		this.capacity = capacity;
 	}
 }