+++ modules/core/src/main/java/io/github/palexdev/mfxcore/utils/fx/CSSFragment.java
@@ -18,8 +18,24 @@
 
 package io.github.palexdev.mfxcore.utils.fx;
 
+import io.github.palexdev.mfxcore.controls.Text;
+import javafx.geometry.HPos;
+import javafx.geometry.Orientation;
+import javafx.geometry.Pos;
+import javafx.geometry.VPos;
+import javafx.scene.Cursor;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.ContentDisplay;
+import javafx.scene.control.OverrunStyle;
+import javafx.scene.effect.BlendMode;
+import javafx.scene.shape.StrokeLineCap;
+import javafx.scene.shape.StrokeLineJoin;
+import javafx.scene.shape.StrokeType;
+import javafx.scene.text.FontPosture;
+import javafx.scene.text.FontSmoothingType;
+import javafx.scene.text.FontWeight;
+import javafx.scene.text.TextAlignment;
 
 import java.util.Base64;
 import java.util.Objects;
@@ -246,6 +262,441 @@ public class CSSFragment {
 		 */
 		public void applyOn(Scene scene) {
 			toCSS().applyOn(scene);
+		}
+
+		//================================================================================
+		// Convenient Methods
+		//================================================================================
+		// NODE
+		public Builder blendMode(BlendMode val) {
+			addStyle("-prop: " + val);
+			return this;
+		}
+
+		public Builder cursor(Cursor val) {
+			addStyle("-fx-cursor: " + val);
+			return this;
+		}
+
+		public Builder focusTraversable(boolean val) {
+			addStyle("-fx-focus-traversable: " + val);
+			return this;
+		}
+
+		public Builder viewOrder(int val) {
+			addStyle("-fx-view-order: " + val);
+			return this;
+		}
+
+		public Builder opacity(double val) {
+			addStyle("-fx-opacity: " + val);
+			return this;
+		}
+
+		public Builder rotate(double val) {
+			addStyle("-fx-rotate: " + val);
+			return this;
+		}
+
+		public Builder scaleX(double val) {
+			addStyle("-fx-scale-x: " + val);
+			return this;
+		}
+
+		public Builder scaleY(double val) {
+			addStyle("-fx-scale-y: " + val);
+			return this;
+		}
+
+		public Builder scaleZ(double val) {
+			addStyle("-fx-scale-z: " + val);
+			return this;
+		}
+
+		public Builder translateX(double val) {
+			addStyle("-fx-translate-x: " + val);
+			return this;
+		}
+
+		public Builder translateY(double val) {
+			addStyle("-fx-translate-y: " + val);
+			return this;
+		}
+
+		public Builder translateZ(double val) {
+			addStyle("-fx-translate-z: " + val);
+			return this;
+		}
+
+		public Builder visibility(String val) {
+			addStyle("visibility: " + val);
+			return this;
+		}
+
+		public Builder managed(boolean val) {
+			addStyle("-fx-managed: " + val);
+			return this;
+		}
+
+		// REGION
+		public Builder background(String val) {
+			addStyle("-fx-background-color: " + val);
+			return this;
+		}
+
+		public Builder backgroundInsets(String val) {
+			addStyle("-fx-background-insets: " + val);
+			return this;
+		}
+
+		public Builder backgroundRadius(String val) {
+			addStyle("-fx-background-radius: " + val);
+			return this;
+		}
+
+		public Builder border(String val) {
+			addStyle("-fx-border-color: " + val);
+			return this;
+		}
+
+		public Builder borderInsets(String val) {
+			addStyle("-fx-border-insets: " + val);
+			return this;
+		}
+
+		public Builder borderRadius(String val) {
+			addStyle("-fx-border-radius: " + val);
+			return this;
+		}
+
+		public Builder padding(String val) {
+			addStyle("-fx-padding: " + val);
+			return this;
+		}
+
+		public Builder positionShape(boolean val) {
+			addStyle("-fx-position-shape: " + val);
+			return this;
+		}
+
+		public Builder scaleShape(boolean val) {
+			addStyle("-fx-scale-shape: " + val);
+			return this;
+		}
+
+		public Builder shape(String val) {
+			addStyle("-fx-shape: " + val);
+			return this;
+		}
+
+		public Builder snapToPixel(boolean val) {
+			addStyle("-fx-snap-to-pixel: " + val);
+			return this;
+		}
+
+		public Builder minHeight(double val) {
+			addStyle("-fx-min-height: " + val);
+			return this;
+		}
+
+		public Builder prefHeight(double val) {
+			addStyle("-fx-pref-height: " + val);
+			return this;
+		}
+
+		public Builder maxHeight(double val) {
+			addStyle("-fx-max-height: " + val);
+			return this;
+		}
+
+		public Builder minWidth(double val) {
+			addStyle("-fx-min-width: " + val);
+			return this;
+		}
+
+		public Builder prefWidth(double val) {
+			addStyle("-fx-pref-width: " + val);
+			return this;
+		}
+
+		public Builder maxWidth(double val) {
+			addStyle("-fx-max-width: " + val);
+			return this;
+		}
+
+		// PANES
+		public Builder hGap(double val) {
+			addStyle("-fx-hgap: " + val);
+			return this;
+		}
+
+		public Builder vGap(double val) {
+			addStyle("-fx-vgap: " + val);
+			return this;
+		}
+
+		public Builder alignment(Pos val) {
+			addStyle("-fx-alignment: " + val);
+			return this;
+		}
+
+		public Builder orientation(Orientation val) {
+			addStyle("-fx-orientation: " + val);
+			return this;
+		}
+
+		public Builder columnHAlignment(HPos val) {
+			addStyle("-fx-column-halignment: " + val);
+			return this;
+		}
+
+		public Builder rowVAlignment(VPos val) {
+			addStyle("-fx-row-valignment: " + val);
+			return this;
+		}
+
+		public Builder gridLineVisible(boolean val) {
+			addStyle("-fx-grid-lines-visible: " + val);
+			return this;
+		}
+
+		public Builder spacing(double val) {
+			addStyle("-fx-spacing: " + val);
+			return this;
+		}
+
+		public Builder fillHeight(boolean val) {
+			addStyle("-fx-fill-height: " + val);
+			return this;
+		}
+
+		public Builder fillWidth(boolean val) {
+			addStyle("-fx-fill-width: " + val);
+			return this;
+		}
+
+		public Builder prefRows(int val) {
+			addStyle("-fx-pref-rows: " + val);
+			return this;
+		}
+
+		public Builder prefColumns(int val) {
+			addStyle("-fx-pref-columns: " + val);
+			return this;
+		}
+
+		public Builder prefTileWidth(double val) {
+			addStyle("-fx-pref-tile-width: " + val);
+			return this;
+		}
+
+		public Builder prefTileHeight(double val) {
+			addStyle("-fx-pref-tile-height: " + val);
+			return this;
+		}
+
+		public Builder tileAlignment(Pos val) {
+			addStyle("-fx-tile-alignment: " + val);
+			return this;
+		}
+
+		// SHAPE
+		public Builder fill(String val) {
+			addStyle("-fx-fill: " + val);
+			return this;
+		}
+
+		public Builder smooth(boolean val) {
+			addStyle("-fx-smooth: " + val);
+			return this;
+		}
+
+		public Builder stroke(String val) {
+			addStyle("-fx-stroke: " + val);
+			return this;
+		}
+
+		public Builder strokeType(StrokeType val) {
+			addStyle("-fx-stroke-type: " + val);
+			return this;
+		}
+
+		public Builder strokeDashArray(String val) {
+			addStyle("-fx-stroke-dash-array: " + val);
+			return this;
+		}
+
+		public Builder strokeDashOffset(double val) {
+			addStyle("-fx-stroke-dash-offset: " + val);
+			return this;
+		}
+
+		public Builder strokeLineCap(StrokeLineCap val) {
+			addStyle("-fx-stroke-line-cap: " + val);
+			return this;
+		}
+
+		public Builder strokeLineJoin(StrokeLineJoin val) {
+			addStyle("-fx-stroke-line-join: " + val);
+			return this;
+		}
+
+		public Builder strokeMiterLimit(double val) {
+			addStyle("-fx-stroke-miter-limit: " + val);
+			return this;
+		}
+
+		public Builder strokeWidth(double val) {
+			addStyle("-fx-stroke-width: " + val);
+			return this;
+		}
+
+		public Builder arcHeight(double val) {
+			addStyle("-fx-arc-height: " + val);
+			return this;
+		}
+
+		public Builder arcWidth(String val) {
+			addStyle("-fx-arc-width: " + val);
+			return this;
+		}
+
+		// TEXT
+		public Builder fontFamily(String val) {
+			addStyle("-fx-font-family: " + val);
+			return this;
+		}
+
+		public Builder fontSize(String val) {
+			addStyle("-fx-font-size: " + val);
+			return this;
+		}
+
+		public Builder fontWeight(FontWeight val) {
+			addStyle("-fx-font-weight: " + val);
+			return this;
+		}
+
+		public Builder fontStyle(FontPosture val) {
+			addStyle("-prop: " + val);
+			return this;
+		}
+
+		public Builder fontSmoothingType(FontSmoothingType val) {
+			addStyle("-fx-font-smoothing-type: " + val);
+			return this;
+		}
+
+		public Builder tabSize(int val) {
+			addStyle("-fx-tab-size: " + val);
+			return this;
+		}
+
+		public Builder textAlignment(TextAlignment val) {
+			addStyle("-fx-text-alignment: " + val);
+			return this;
+		}
+
+		public Builder textOrigin(VPos val) {
+			addStyle("-fx-text-origin: " + val);
+			return this;
+		}
+
+		public Builder underline(boolean val) {
+			addStyle("-fx-underline: " + val);
+			return this;
+		}
+
+		/**
+		 * This will work only for {@link Text}.
+		 */
+		public Builder wrappingWidth(double val) {
+			addStyle("-fx-wrapping-width: " + val);
+			return this;
+		}
+
+		// LABEL
+		public Builder textOverrun(OverrunStyle val) {
+			addStyle("-fx-text-overrun: " + val);
+			return this;
+		}
+
+		public Builder wrapText(boolean val) {
+			addStyle("-fx-wrap-text: " + val);
+			return this;
+		}
+
+		public Builder contentDisplay(ContentDisplay val) {
+			addStyle("-fx-content-display: " + val);
+			return this;
+		}
+
+		public Builder graphicTextGap(double val) {
+			addStyle("-fx-graphic-text-gap: " + val);
+			return this;
+		}
+
+		public Builder labelPadding(double val) {
+			addStyle("-fx-label-padding: " + val);
+			return this;
+		}
+
+		public Builder textFill(String val) {
+			addStyle("-fx-text-fill: " + val);
+			return this;
+		}
+
+		public Builder ellipsisString(String val) {
+			addStyle("-fx-ellipsis-string: " + val);
+			return this;
+		}
+
+		// TEXT INPUT CONTROL
+		public Builder promptTextFill(String val) {
+			addStyle("-fx-prompt-text-fill: " + val);
+			return this;
+		}
+
+		public Builder highlightFill(String val) {
+			addStyle("-fx-highlight-fill: " + val);
+			return this;
+		}
+
+		public Builder highlightTextFill(String val) {
+			addStyle("-fx-highlight-text-fill: " + val);
+			return this;
+		}
+
+		public Builder displayCaret(boolean val) {
+			addStyle("-fx-display-caret: " + val);
+			return this;
+		}
+
+		// TEXT AREA
+		public Builder prefColumnCount(int val) {
+			addStyle("-fx-pref-column-count: " + val);
+			return this;
+		}
+
+		public Builder prefRowCount(int val) {
+			addStyle("-fx-pref-row-count: " + val);
+			return this;
+		}
+
+		// WEB VIEW
+		public Builder contextMenuEnabled(boolean val) {
+			addStyle("-fx-context-menu-enabled: " + val);
+			return this;
+		}
+
+		public Builder pageFill(String val) {
+			addStyle("-fx-page-fill: " + val);
+			return this;
+		}
+
+		public Builder fontScale(double val) {
+			addStyle("-fx-font-scale: " + val);
+			return this;
 		}
 	}
 }